Move over to the GPL code base.
Some files get lost. They're not ones we care about.

+++ sys/src/ape/lib/bsd/ntohl.c
@@ -1,5 +1,14 @@
+/* 
+ * This file is part of the UCB release of Plan 9. It is subject to the license
+ * terms in the LICENSE file found in the top-level directory of this
+ * distribution and at http://akaros.cs.berkeley.edu/files/Plan9License. No
+ * part of the UCB release of Plan 9, including this file, may be copied,
+ * modified, propagated, or distributed except according to the terms contained
+ * in the LICENSE file.
+ */
+
 unsigned long
-ntohl(int x)
+ntohl(unsigned long x)
 {
 	unsigned long n;
 	unsigned char *p;
@@ -24,7 +33,7 @@ htonl(unsigned long h)
 }
 
 unsigned short
-ntohs(int x)
+ntohs(unsigned short x)
 {
 	unsigned short n;
 	unsigned char *p;
